Configure and build before 'cabal test' and 'cabal bench'.
The test and benchmark cabal-install commands now accept ConfigFlags and
ConfigExFlags in the same way as the install command, so that the
package can be configured before running test suites or benchmarks. Also
like the install command, the test and bench commands run the build
command with the default options to perform any necessary rebuilds.