Build detailed test suite pieces with own ComponentLocalBuildInfo.
Bug #1287. Type 'detailed-0.9' test suites were failing to build because the
ComponentLocalBuildInfo for the test suite was being used to build the library
and stub executable. The test suite's ComponentLocalBuildInfo should only be used
to create the CLBIs for the library and stub executable, as they have their own
CLBI constructors.