Build detailed test suites as executables directly
The old method of building detailed test suites was:
1. Build the test suite as a library with one exposed module.
2. In-place register the library in its own package database.
3. Build a wrapper that depends on the library.

This is a very fragile way of doing things. Naming is a particularly bad
issue: we must be careful that names of various components (the test
suite library, the test suite wrapper, other executables, the package
itself, etc.) do not collide. We've been lucky so far, but I can't even
speculate what would happen if someone intentionally misnamed
components. A worse restriction on naming is that detailed test suite
names must not contain digits, or the name generated for the test suite
library will be unparseable!

This is also a very arcane way of doing things. We built this interface
envisioning that a diverse ecosystem of test platforms would develop,
including different test runners that would link against the test suite
libraries and display the results in new, exciting ways. That never
happened, probably because the process of finding each test suite's
library and in-place package database and tricking Cabal into building
your test wrapper is, to be charitable, arcane. I'm not sure I could
tell you how to do that reliably, and I wrote the code!

The new method of building test suites is:
1. Build an executable where 'main-is' points to the source code of the
test wrapper and 'other-modules' includes all modules from the detailed
test suite. The rest of the BuildInfo should be copied from the detailed
test suite stanza.

Most importantly, there's no name fragility here: as long as the test
suite doesn't share a name with any other component, nothing should go
wrong. (That's already a standard part of 'cabal check'.) Implementing
custom test runners should also be substantially simpler: it's just a
matter of swapping out Cabal's wrapper code. (More on that point in
another commit.)