D.S.Utils: add withRawSystemIOEnv
Unlike rawSystemIOWithEnv, which blocks waiting for the child process to
finish, withRawSystemIOEnv takes an IO action which is performed after
the child is spawned but before blocking to wait for the exit code. This
allows us to, e.g., stream the output of a child process to stdout while
the process runs in the background.